Rename jasper restart -> restart jasper
Fix amboy job tests and scheduler wrapper tests
