+++ units/crons.go
@@ -845,9 +845,9 @@ func PopulateHostSetupJobs(env evergreen.Environment) amboy.QueueOperation {
 	}
 }
 
-// PopulateHostJasperRestartJobs enqueues the jobs to restart the Jasper service
+// PopulateHostRestartJasperJobs enqueues the jobs to restart the Jasper service
 // on the host.
-func PopulateHostJasperRestartJobs(env evergreen.Environment) amboy.QueueOperation {
+func PopulateHostRestartJasperJobs(env evergreen.Environment) amboy.QueueOperation {
 	return func(ctx context.Context, queue amboy.Queue) error {
 		flags, err := evergreen.GetServiceFlags()
 		if err != nil {
@@ -863,11 +863,11 @@ func PopulateHostJasperRestartJobs(env evergreen.Environment) amboy.QueueOperati
 			return nil
 		}
 
-		hosts, err := host.FindByNeedsJasperRestart()
+		hosts, err := host.FindByNeedsToRestartJasper()
 		if err != nil {
 			grip.Error(message.WrapError(err, message.Fields{
 				"operation": "Jasper service restart",
-				"cron":      jasperRestartJobName,
+				"cron":      restartJasperJobName,
 				"impact":    "existing hosts will not have their Jasper services restarted",
 			}))
 			return errors.Wrap(err, "problem finding hosts that need their Jasper service restarted")
