+++ units/crons_remote_five_minute.go
@@ -50,7 +50,6 @@ func (j *cronsRemoteFiveMinuteJob) Run(ctx context.Context) {
 	ops := []amboy.QueueOperation{
 		PopulateTaskMonitoring(5),
 		PopulateActivationJobs(10),
-		PopulateHostJasperRestartJobs(j.env),
 	}
 
 	queue := j.env.RemoteQueue()
