+++ units/crons_remote_hour.go
@@ -50,6 +50,7 @@ func (j *cronsRemoteHourJob) Run(ctx context.Context) {
 	ops := []amboy.QueueOperation{
 		PopulateCacheHistoricalTestDataJob(2),
 		PopulateHostProvisioningConversionJobs(j.env),
+		PopulateHostRestartJasperJobs(j.env),
 		PopulateSpawnhostExpirationCheckJob(),
 		PopulateVolumeExpirationCheckJob(),
 		PopulateVolumeExpirationJob(),
